Disable MySQL binary logs.
In the old MySQL version, they *were* disabled.  In the new one (at least
in the supplied configuration file) they were enabled, without auto
expiration. This means that *all* data transactions are stored for
eternity.  Your database may be empty (thus small), but the binary logs
could restore your database as it existed at any point in time.  With lots
and lots of data moving in and out, the binary logs take up many gigabytes
of storage.  Of course, higher trigger rates (say, KASCADE), will make
this worse.  And so KASCADE ran out of disk space in a week.

Problem: disabling the binary logs does not remove them.  So let's get
this quickly out the door and in a future update, search for and remove
all binary logs...