Changed to always delete events that are to be discarded.
Previously it was the case that bad events (those that fail in the interpreter, not those that cannot be processed because the storage is full) were only deleted once there are no more good events. If the incidence of good events is high enough this can cause bad events to retried over and over again to no avail, each time printing error messages.

Now events that can be discarded are deleted immediately instead of waiting until there are no more good events.