Actually enable the delay for starting the Monitor
This delay was added to prevent the Monitor from starting to soon. MySQL
should be ready to accept connections when the Monitor starts. Due to
mixing
spaces and tabs in earlier versions this command is now run as
expected. It
was only executed when there was an error starting the Weather DAQ. I
reduced
the delay because it seems that it has been working fine without it.