Send the low/high thresholds for new DAQ as absolute ADC
Use the mv_to_adc and adc_baseline values to calculate the thresholds
as ADC.
This would make it easier for later analysis to know the exact value
that was used.
Absolute ADC values are positive instead of negative like the mV,
this way you can determine the units of the value.