Increase NRDP POST timeout
Increase timeout to 5 seconds to avoid timeouts during startup.

During startup of the hsmonitor, I have seen some requests timeout.
This throws a few warnings (one for each passive check/NRDP POST)
which can be disregared, but may perhaps result in bug reports
and/or support requests. Increase the timeout to prevent this.