hsmonitor: Show HTTP exception only in debug log
In the hsmonitor the loglevel "on screen" is usually set to INFO.
The loglevel on disk is set to DEBUG.

Print the full `requests.exception` only in the debug log. So only
relevant warning on screen. This will hopefully prevent bug reports
and/or support requests when there are (non permanent) connection
error that can be disregarded.