Fix TriggerTime offset between GPS and PC clock
The DAQ determines the offset between GPS time and the local PC clock.
This offset is used in the hsmonitor to compute the time off the last trigger.

Between versions v3 and v4 of the DAQ the SIGN of the offset changed!

DAQ v4: offset = GPS time - PC clock
(The number of seconds GPS is ahead of the PC clock).
This fixes the sign change.